# fixes #504 fread now handles all kind of NAs without coercion to char