… man/fread.Rd