Tests added.  Test 882 adjusted to match correct behaviour. 
I tested this a lot (also `test.data.table()` helps! ), but may be we need to add more tests for this. I believe I need some help with writing good tests.

Now `na.strings = NULL` means no coercion at all. Old hardcoded handling for "NA" string removed from everywhere.
@arunsrinivasan, please review code before merging. This PR touches a lot of `fread.c` internals. It affects `Strto*()` functions which are workhorses for data reading and conversion.
Also It would be great if @mattdowle also will take a look, because as I see he wrote most of the `fread` code.