@dselivanov 2015-07-21
strange behaviour - CI tests 908, 1343, 1344 were failed, however all tests were passed on my machine. I'll check.