@jangorecki 2015-07-21
It happens sometimes to fail the tests while on the same code next built will pass them. In that case, all mentioned tests (908, 1343, 1344) are related to fread.  
I did run the the test against your branch and it failed on my machine, it may be related to OS.
Reproducible script below.
```r
library(devtools)
install_github("dselivanov/data.table@fread_improvements")
library(data.table)
test <- function(num, x, y) all.equal(x,y)
test(908, fread("A,B,C\n1,3,\n2,4,\n"), data.table(A=1:2,B=3:4,C=NA)) # where NA is type logical
# [1] "Component “C”: Modes: numeric, logical"               "Component “C”: target is numeric, current is logical"
test(1343, fread("A,B\n1,TRUE\n2,\n3,F"), data.table(A=1:3, B=c(TRUE,NA,FALSE)))
# [1] "Component “B”: Modes: character, logical"               "Component “B”: target is character, current is logical"
test(1344, fread("A,B\n1,T\n2,NA\n3,"), data.table(A=1:3, B=c(TRUE,NA,NA)))
# [1] "Component “B”: Modes: character, logical"               "Component “B”: target is character, current is logical"
# Warning message:
# In fread("A,B\n1,T\n2,NA\n3,") :
#   Bumped column 2 to type character on data row 1, field contains 'T'. Coercing previously read values in this column from logical, integer or numeric back to character which may not be lossless; e.g., if '00' and '000' occurred before they will now be just '0', and there may be inconsistencies with treatment of ',,' and ',NA,' too (if they occurred in this column before the bump). If this matters please rerun and set 'colClasses' to 'character' for this column. Please note that column type detection uses the first 5 rows, the middle 5 rows and the last 5 rows, so hopefully this message should be very rare. If reporting to datatable-help, please rerun and include the output from verbose=TRUE.
```

```
R version 3.2.1 (2015-06-18)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 14.04.2 LTS

locale:
 [1] LC_CTYPE=en_GB.UTF-8       LC_NUMERIC=C               LC_TIME=en_GB.UTF-8        LC_COLLATE=en_GB.UTF-8     LC_MONETARY=en_GB.UTF-8    LC_MESSAGES=en_GB.UTF-8   
 [7] LC_PAPER=en_GB.UTF-8       LC_NAME=C                  LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=en_GB.UTF-8 LC_IDENTIFICATION=C       

attached base packages:
[1] stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
[1] data.table_1.9.5 devtools_1.8.0  

loaded via a namespace (and not attached):
 [1] httr_1.0.0      R6_2.1.0        magrittr_1.5    rversions_1.0.2 tools_3.2.1     curl_0.9.1      Rcpp_0.11.6     memoise_0.2.1   xml2_0.1.1      stringi_0.5-5  
[11] knitr_1.10.5    git2r_0.10.1    stringr_1.0.0   digest_0.6.8    chron_2.3-47   
```