@arunsrinivasan 2015-07-21
@dselivanov thanks. I'll leave it to @mattdowle for this one.

What'd be very useful is to run a sufficiently large file (both with, and without using this argument) using `fread()` before and after your fix and report the timings.