@dselivanov 2015-07-21
@jangorecki, I'm on mac os
> R version 3.2.1 (2015-06-18)
Platform: x86_64-apple-darwin13.4.0 (64-bit)
Running under: OS X 10.10.4 (Yosemite)
locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8
attached base packages:
[1] stats     graphics  grDevices utils     datasets  methods   base     
other attached packages:
[1] data.table_1.9.5
loaded via a namespace (and not attached):
[1] tools_3.2.1  chron_2.3-47

But I'll try these test on ubuntu.

@arunsrinivasan Seems, timings for master and fread_improvements branches are quite similar.

Benchmark:
**generate data**
```r
library(data.table)
K <- 1e7
DT <- data.table(int = 1:K, 
                 char = sample(letters, size = K, replace = T), 
                 float = 1:K + 0.1, 
                 bool = sample( c(T, F), K, replace = T))

fractions <- c(0.001, 0.01, 0.1, 0.3)
files <- paste0("~/fr_", fractions)
for( fr in seq_along(fractions) ) {
  DT_NA <- copy(DT)
  for (j in seq_len( ncol(DT) )) {
    i <- sample(K, ceiling(fractions[[fr]] * K) )
    set(x = DT_NA, i = i, j = j, value = NA)
  }
  write.table(DT_NA, file = files[[fr]], quote = F, sep = ',', row.names = F, col.names = T)
}
```
**NEW fread branch**
```r
library(devtools)
install_github("dselivanov/data.table@fread_improvements")
library(data.table)
fractions <- c(0.001, 0.01, 0.1, 0.3)
files <- paste0("~/fr_", fractions)
for (f in files) {
  # this copy prevents reading from already mmaped file?
  f1 <- '~/tempfile'
  system(paste('cp', f, f1))
  #*****************************************************
  timing <- system.time(DT_fread <- fread(f1, na.strings = "NA", verbose = F))
  print(f)
  print(timing)
}
```
>[1] "~/fr_0.001"
   user  system elapsed 
  1.546   0.146   1.692 
[1] "~/fr_0.01"
   user  system elapsed 
  1.554   0.156   1.708 
[1] "~/fr_0.1"
   user  system elapsed 
  1.559   0.104   1.663 
[1] "~/fr_0.3"
   user  system elapsed 
  1.823   0.086   1.910 

**OLD fread branch**
```r
library(devtools)
install_github('rdatatable/data.table')
library(data.table)
fractions <- c(0.001, 0.01, 0.1, 0.3)
files <- paste0("~/fr_", fractions)
for (f in files) {
  # this copy prevents reading from already mmaped file?
  f2 <- '~/tempfile'
  system(paste('cp', f, f2))
  #*****************************************************
  system(paste('cp', f, f2))
  timing <- system.time(DT_fread <- fread(f2, na.strings = "NA", verbose = F))
  print(f)
  print(timing)
}
```
>[1] "~/fr_0.001"
   user  system elapsed 
  1.456   0.149   1.605 
[1] "~/fr_0.01"
   user  system elapsed 
  1.586   0.137   1.724 
[1] "~/fr_0.1"
   user  system elapsed 
  1.490   0.092   1.582 
[1] "~/fr_0.3"
   user  system elapsed 
  1.528   0.088   1.617