@dselivanov 2015-07-22
I fixed - variable `na_found_flag ` was uninitialised. Compile with `-pedantic` flag helps to locate issue.  

But how do you think, guys, is it right expected behaviour, that in these cases empty string at the end of line converted into `NA`s, even if we don't set `na.strings=""`?