@arunsrinivasan 2015-08-18
@dselivanov thanks for your time. I'll have time this weekend to take a look at this. I don't foresee any issues.. so it's very likely I'll merge this. Thanks for the benchmarks and code to reproduce it. I ran it on 1e8 rows, and things weren't much different. That's great!